Preparation for future implementation: infrastructure is now in place to detect physically identical but programmatically distinct diagrams.

+++ Model.cpp
@@ -87,7 +87,7 @@ loopdict getLoopDictionary(const n1dict& inters)
 		for (int num_incoming{ 1 }; num_incoming < static_cast<int>(s); ++num_incoming) {
 			//We need to select num_incoming-1 particles to move from the left hand of the interaction to the right
 			//Set up a bitmask with num_incoming-1 1s and then enough 0s to make up the slack
-			std::string bitmask(num_incoming - 1, '1');
+			std::string bitmask((long) num_incoming - 1, '1');
 			bitmask.resize(s, '0');
 			do {
 				//Keep the current output particle as a new input particle
